refactor: Auth page refactor to react-query, fix: Layout and header error fix
