+++ src/components/Common/Layout/index.tsx
@@ -7,6 +7,7 @@ const Layout: FC<LayoutProps> = ({
   children,
   hasHeader = false,
   hasTapBar = false,
+  showMenuButton = false,
 }: LayoutProps) => {
   const headerHeight = 73; // 헤더의 높이
   const tapBarHeight = 73; // TapBar의 높이
@@ -30,7 +31,7 @@ const Layout: FC<LayoutProps> = ({
           <MainPageHeader
             title="깊은산 골짜기"
             fontFamily="Cafe24Ssurround"
-            showMenuButton={true}
+            showMenuButton={showMenuButton}
           />
         </Box>
       )}
