+++ src/pages/Auth/Register/index.tsx
@@ -1,15 +1,11 @@
 import { FC } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { Formik, Field } from 'formik';
 import CustomButton from '../../../components/Common/CustomButton';
 import Layout from '../../../components/Common/Layout';
 import Logo from '../../../assets/images/Logo.png';
 import { buttonStyle } from '../../../styles/customChakraPropsStyle';
 import { emailRegEx, passwordRegEx } from '../../../utils/Regex';
-import { register } from '../../../api/Auth/AuthService';
-import useErrorToast from '../../../hooks/useErrorToast';
-import useSuccessToast from '../../../hooks/useSuccessToast';
-import { useMutation } from '@tanstack/react-query';
+import useRegisterMutation from '../../../queries/useRegisterMutation';
 import {
   Flex,
   FormControl,
@@ -20,48 +16,9 @@ import {
   Link,
   Text,
 } from '@chakra-ui/react';
-import {
-  잘못된요청,
-  이미존재하는이메일입니다,
-  이메일이중복되었어요,
-  닉네임이중복되었어요,
-  이미존재하는닉네임입니다,
-  서버오류,
-} from '../../../constant/constant';
 
 const Register: FC = () => {
-  const navigate = useNavigate();
-  const { errorToast } = useErrorToast();
-  const { successToast } = useSuccessToast();
-
-  const mutation = useMutation({
-    mutationFn: register,
-    onSuccess: (response) => {
-      localStorage.setItem('token', response.data.access_token);
-      navigate('/');
-      successToast({
-        title: '회원가입 성공!',
-        description: '회원가입에 성공하였습니다.',
-      });
-    },
-    onError: (err: any) => {
-      if (err.response.status === 400) {
-        errorToast(잘못된요청);
-      }
-      if (err.response.status === 409) {
-        if (err.response.data === 이메일이중복되었어요) {
-          errorToast(이미존재하는이메일입니다);
-        }
-        if (err.response.data === 닉네임이중복되었어요) {
-          errorToast(이미존재하는닉네임입니다);
-        }
-      }
-      if (err.response.status === 500) {
-        errorToast(서버오류);
-      }
-    },
-  });
-
+  const mutation = useRegisterMutation();
   const handleSubmit = (values: {
     login_email: string;
     name: string;
